llm-core/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
requests